refactor: remove schema to a separate file from form
